Перенесла навигатор в отдельный компонент
