+++ project/src/components/favorites/favorites.tsx
@@ -1,5 +1,6 @@
 import { Link } from 'react-router-dom';
 
+import ConnectedHeaderNav from '../header-nav/header-nav';
 import FavoritePlaceCard from '../favorite-place-card/favorite-place-card';
 import Logo from '../logo/logo';
 
@@ -34,22 +35,7 @@ function Favorites({offers, favorites, onFavoritesClick}: FavoritesProps): JSX.E
             <div className="header__left">
               <Logo />
             </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to="/">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="/">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
+            <ConnectedHeaderNav />
           </div>
         </div>
       </header>
