+++ project/src/components/header-nav/header-nav.tsx
@@ -1,0 +1,63 @@
+import { connect, ConnectedProps } from 'react-redux';
+import { Link } from 'react-router-dom';
+
+import { AppRoute, AuthorizationStatus } from '../../const';
+import { logoutAction } from '../../store/api-actions';
+import { State } from '../../types/state';
+import { ThunkAppDispatch } from '../../types/action';
+
+const mapStateToProps = ({authorizationStatus, authInfo}: State) => ({
+  authorizationStatus,
+  authInfo,
+});
+
+const mapDispatchToProps = (dispatch: ThunkAppDispatch) => ({
+  onLogout() {
+    dispatch(logoutAction());
+  },
+});
+
+const connector = connect(mapStateToProps, mapDispatchToProps);
+
+type PropsFromRedux = ConnectedProps<typeof connector>;
+
+function HeaderNav(props: PropsFromRedux): JSX.Element {
+  const {
+    authorizationStatus,
+    authInfo,
+    onLogout,
+  } = props;
+  return (
+    <nav className="header__nav">
+      <ul className="header__nav-list">
+        {authorizationStatus === AuthorizationStatus.Auth ? (
+          <>
+            <li className="header__nav-item user">
+              <Link className="header__nav-link header__nav-link--profile" to={AppRoute.Favorites}>
+                <div className="header__avatar-wrapper user__avatar-wrapper">
+                </div>
+                <span className="header__user-name user__name">{authInfo?.email}</span>
+              </Link>
+            </li>
+            <li className="header__nav-item">
+              <Link onClick={onLogout} className="header__nav-link" to="/">
+                <span className="header__signout">Sign out</span>
+              </Link>
+            </li>
+          </>
+        ) : (
+          <li className="header__nav-item user">
+            <Link className="header__nav-link header__nav-link--profile" to={AppRoute.SignIn}>
+              <div className="header__avatar-wrapper user__avatar-wrapper">
+              </div>
+              <span className="header__login">Sign in</span>
+            </Link>
+          </li>
+        )}
+      </ul>
+    </nav>
+  );
+}
+
+export {HeaderNav};
+export default connector(HeaderNav);
