+++ project/src/components/property/property.tsx
@@ -1,8 +1,8 @@
 import { connect, ConnectedProps } from 'react-redux';
 import { useEffect } from 'react';
-import { Link } from 'react-router-dom';
 import { useParams } from 'react-router';
 
+import ConnectedHeaderNav from '../header-nav/header-nav';
 import ConnectedMap from '../map/map';
 import Logo from '../logo/logo';
 import NotFound from '../not-found/not-found';
@@ -64,22 +64,7 @@ function Property(props: ConnectedComponentProps): JSX.Element {
             <div className="header__left">
               <Logo />
             </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to="/">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="/">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
+            <ConnectedHeaderNav />
           </div>
         </div>
       </header>
